auth_service/requirements.txt
flask
flask-restx
python-dotenv
pytest
pytest-cov
PyJWT